+++ packages/expo-modules-core/ios/Arguments/AnyArgument.swift
@@ -3,10 +3,15 @@
 /**
  A protocol for classes/structs accepted as an argument of functions.
  */
-public protocol AnyArgument {}
+public protocol AnyArgument {
+  static func getDynamicType() -> AnyDynamicType
+}
 
-// Extend the optional type - this is required to support optional arguments.
-extension Optional: AnyArgument where Wrapped: AnyArgument {}
+extension AnyArgument {
+  public static func getDynamicType() -> AnyDynamicType {
+    return DynamicRawType(innerType: Self.self)
+  }
+}
 
 // Extend the primitive types — these may come from React Native bridge.
 extension Bool: AnyArgument {}
@@ -28,6 +33,16 @@ extension Double: AnyArgument {}
 extension CGFloat: AnyArgument {}
 
 extension String: AnyArgument {}
-extension Array: AnyArgument {}
 extension Dictionary: AnyArgument {}
 
+extension Optional: AnyArgument where Wrapped: AnyArgument {
+  public static func getDynamicType() -> AnyDynamicType {
+    return DynamicOptionalType(wrappedType: ~Wrapped.self)
+  }
+}
+
+extension Array: AnyArgument {
+  public static func getDynamicType() -> AnyDynamicType {
+    return DynamicArrayType(elementType: ~Element.self)
+  }
+}